Gallery: Move to 'General' category
